requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0
openai==1.3.9
SQLAlchemy==2.0.25
psycopg2==2.9.9
python-dotenv==1.0.1
